pyproject.toml
[project]
name = "automated-youtube-mixtape-creation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.135.1",
    "ffmpeg>=1.4",
    "ffmpeg-python>=0.2.0",
    "ipykernel>=7.2.0",
    "librosa>=0.11.0",
    "matplotlib>=3.10.8",
    "moviepy>=2.2.1",
    "numpy>=2.4.2",
    "pillow>=11.3.0",
    "pydub>=0.25.1",
    "requests>=2.32.5",
    "python-multipart>=0.0.22",
    "scipy>=1.17.1",
    "streamlit>=1.55.0",
    "uvicorn[standard]>=0.41.0",
]
